+++ apps/app/.env.example
@@ -47,6 +47,8 @@ NEXT_PUBLIC_INTERCOM_APP_ID=""
 NEXT_PUBLIC_SITE_PASSWORD=""
 NEXT_PUBLIC_HUBSPOT_PORTAL_ID=""
 NEXT_PUBLIC_HUBSPOT_FORM_ID=""
+NEXT_PUBLIC_HOTJAR_ID=""
+NEXT_PUBLIC_GOOGLE_ANALYTICS_ID=""
 
 GCP_BUCKET_NAME=livepeer-clips-staging
 GCP_CREDENTIALS="
