+++ apps/app/app/layout.tsx
@@ -6,7 +6,7 @@ import Intercom from "@/components/intercom";
 import { Metadata } from "next";
 import { MixpanelProvider } from "@/components/analytics/MixpanelProvider";
 import { Inter, Playfair, Open_Sans } from "next/font/google";
-import Script from "next/script";
+import { ThirdPartyAnalytics } from "@/scripts/analytics";
 
 const playfair = Playfair({
   variable: "--font-playfair",
@@ -54,44 +54,7 @@ const RootLayout = ({ children }: RootLayoutProperties) => (
           {/* TODO: REENABLE WHEN SHIH-YU IS READY FOR IT <Intercom /> */}
         </DesignSystemProvider>
       </MixpanelProvider>
-      {process.env.NODE_ENV === "production" && (
-        <>
-          {/* Hotjar */}
-          <Script
-            id="contentsquare"
-            strategy="afterInteractive"
-            dangerouslySetInnerHTML={{
-              __html: `
-              (function (c, s, q, u, a, r, e) {
-                  c.hj=c.hj||function(){(c.hj.q=c.hj.q||[]).push(arguments)};
-                  c._hjSettings = { hjid: a };
-                  r = s.getElementsByTagName('head')[0];
-                  e = s.createElement('script'); e.async = true; e.src = q + c._hjSettings.hjid + u; r.appendChild(e);
-              })(window, document, 'https://static.hj.contentsquare.net/c/csq-', '.js', 6381273);
-            `,
-            }}
-          />
-          {/* Google Analytics */}
-          <Script
-            strategy="afterInteractive"
-            src={`https://www.googletagmanager.com/gtag/js?id=G-F7BNWCLKW9`}
-          />
-          <Script
-            id="google-analytics"
-            strategy="afterInteractive"
-            dangerouslySetInnerHTML={{
-              __html: `
-                window.dataLayer = window.dataLayer || [];
-                function gtag(){dataLayer.push(arguments);}
-                gtag('js', new Date());
-                gtag('config', 'G-F7BNWCLKW9', {
-                  page_path: window.location.pathname,
-                });
-              `,
-            }}
-          />
-        </>
-      )}
+      <ThirdPartyAnalytics />
     </body>
   </html>
 );
